Explains Multinomial Logistic Regressions
Explains about how we can use multiple features to predict an output. Data is Cardiotocographic.csv which can be downloaded from GitHub, we also see what is and how we can calculate misclassification error and using z test, how each feature is telling us whether it s playing a significant role in the model or not.